pyproject.toml
[project]
name = "axon-otc-v7"
version = "0.1.0"
description = "Python SDK and CLI for the Axon OTC V7 marketplace."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "web3>=7.0.0",
  "eth-account>=0.13.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
